requirements.txt
Flask==3.0.2
Pillow==10.2.0
matplotlib==3.8.2
numpy==1.26.3